Moved components in single files out of their folders
