+++ components/AnimatedLogo.tsx
